// src/models/Estabelecimentos.ts
export class Estabelecimentos {
  id: number;
  cnpj_cpf: string;
  razao_social: string;
  nome_fantasia: string;
  endereco: string;
  contato: string;
  bairro: string;
  municipio: string;
  email: string;
  tipo_de_risco: string;
  grau_de_risco: string;
  createdAt: Date;
  updatedAt: Date;
  uf: string;
}